Move CycleBuilder inside the CPU
Fixes #147.
