+++ src/main/java/org/edumips64/utils/CycleElement.java
@@ -21,7 +21,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
-package org.edumips64.ui.common;
+package org.edumips64.utils;
 
 import java.util.*;
 import java.util.logging.Logger;
